Add files via upload
Functioning MPPT Integration with the DC-DC Converter. This is with a load of 695 ohms, allowing the PV module power to get to 358 W, but only allows the output power across the load of 280 W. This results in a 78W difference, thus dictating only a 78.2% efficiency. This could be due to the resistors I had to put in front of some capacitors and inductors so matlab wouldn't freak out. It could also be due to the fact that the model is using two pv modules in series instead of one. Additionally it could be due to the mismatch in switch parameters or transformer parameters.