PI Values
Decreased the P and I values, while keeping the same ratio. For example for the first PI controller we had a P value of 1 and a I value of 20 giving us a ratio of 200. I decreased it to have a P value of 0.003125 and an I value of 0.625. For the second PI controller I changed the P value from 100 to 1 and the I value from 1000 to 10. The sine wave actually looks like a "sexy" sine wave with minimal odd harmonics. Don't know why this changed anything, but it did..